Actualización 11 de enero de 2022

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserEditRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
@@ -17,7 +18,8 @@ class UserController extends Controller
 
     public function create()
     {
-        return view('back.users.create');
+        $roles = Role::all()->pluck('name', 'id');
+        return view('back.users.create', compact('roles'));
     }
 
     public function store(UserCreateRequest $request)
@@ -27,18 +29,24 @@ class UserController extends Controller
             + [
                 'password' => bcrypt($request->input('password')),
             ]);
+
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'Usuario ingresado exitosamente.');
     }
 
     public function show(User $user) // Model Binding
     {
         // $user = User::findOrFail($id);
+        $user->load('roles');
         return view('back.users.show', compact('user'));
     }
 
     public function edit(User $user)
     {
-        return view('back.users.edit', compact('user'));
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
+        return view('back.users.edit', compact('user', 'roles'));
     }
 
     public function update(UserEditRequest $request, User $user)
@@ -50,11 +58,16 @@ class UserController extends Controller
         if($password)
             $data['password'] = bcrypt($password);
         $user->update($data);
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
         return redirect()->route('users.show', $user->id)->with('success', 'Usuario actualizado exitosamente.');
     }
 
     public function destroy(User $user)
     {
+        if(auth()->user()->id == $user->id) {
+            return redirect()->route('users');
+        }
         $user->delete();
         return redirect()->route('users')->with('success', 'Usuario eliminado correctamente.');
     }
